requirements.txt
pyrogram
python-dotenv
requests
asyncio